Changed service tests to use JEST
