+++ package.json
@@ -1,43 +1,46 @@
 {
-    "name": "sep-backend",
-    "version": "1.0.0",
-    "description": "Backend for a diabetes platform.",
-    "main": "src\\app.ts",
-    "scripts": {
-        "test-api": "ts-node src/gb/test.ts",
-        "test-server": "ts-node src/server.ts",
-        "test-services": "ts-node src/services/testService.ts",
-        "test": "echo \"Error: no test specified\" && exit 1",
-        "unit-test-services": "ts-node test/services/testServices.ts",
-        "build": "tsc"
-    },
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/RoelMK/SEP-backend.git"
-    },
-    "author": "Group 3",
-    "license": "ISC",
-    "bugs": {
-        "url": "https://github.com/RoelMK/SEP-backend/issues"
-    },
-    "homepage": "https://github.com/RoelMK/SEP-backend#readme",
-    "dependencies": {
-        "axios": "^0.21.1",
-        "date-fns": "^2.21.1",
-        "dotenv": "^8.2.0",
-        "errorhandler": "^1.5.1",
-        "express": "^4.17.1",
-        "fast-csv": "^4.3.6",
-        "fs": "*",
-        "http": "*",
-        "multer": "^1.4.2",
-        "papaparse": "^5.3.0"
-    },
-    "devDependencies": {
-        "@types/node": "^15.0.0",
-        "@types/papaparse": "^5.2.5",
-        "ts-node": "^9.1.1",
-        "tsc": "^1.20150623.0",
-        "typescript": "^4.2.4"
-    }
+  "name": "sep-backend",
+  "version": "1.0.0",
+  "description": "Backend for a diabetes platform.",
+  "main": "src\\app.ts",
+  "scripts": {
+    "test-api": "ts-node src/gb/test.ts",
+    "test-server": "ts-node src/server.ts",
+    "test-services": "ts-node src/services/testService.ts",
+    "test": "jest",
+    "unit-test-services": "ts-node test/services/testServices.ts",
+    "build": "tsc"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/RoelMK/SEP-backend.git"
+  },
+  "author": "Group 3",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/RoelMK/SEP-backend/issues"
+  },
+  "homepage": "https://github.com/RoelMK/SEP-backend#readme",
+  "dependencies": {
+    "axios": "^0.21.1",
+    "date-fns": "^2.21.1",
+    "dotenv": "^8.2.0",
+    "errorhandler": "^1.5.1",
+    "express": "^4.17.1",
+    "fast-csv": "^4.3.6",
+    "fs": "*",
+    "http": "*",
+    "multer": "^1.4.2",
+    "papaparse": "^5.3.0"
+  },
+  "devDependencies": {
+    "@types/jest": "^26.0.23",
+    "@types/node": "^15.0.0",
+    "@types/papaparse": "^5.2.5",
+    "jest": "^26.6.3",
+    "ts-jest": "^26.5.5",
+    "ts-node": "^9.1.1",
+    "tsc": "^1.20150623.0",
+    "typescript": "^4.2.4"
+  }
 }
